requirements.txt
aiogram==3.2.0
alembic==1.13.0
asyncpg==0.29.0
email-validator==2.1.0.post1
fastapi==0.104.1
fluent.runtime==0.4.0
fluent.syntax==0.19.0
orjson==3.9.10
psycopg2-binary==2.9.9
pydantic-settings==2.1.0
redis==5.0.1
SQLAlchemy==2.0.23
ujson==5.8.0
uvicorn==0.24.0.post1
